Day 1, problem 2
I did a first program on time that passed the example input but
failed with real input.

I tried again but is was too late and I needed to sleep.  After that I
cheated and looked for ideas in Reddit.  I found this technique of
reverse the strings easier than the loops backwards that I tried the
first night.